test(cli): split onboard debug suite
